potential bug fix

+++ pyseter/extract.py
@@ -1,6 +1,6 @@
 """Extract features from images with AnyDorsal.
 
-Description 
+Description
 
 Typical usage example:
 
@@ -14,7 +14,7 @@ import os
 from huggingface_hub import hf_hub_download
 from sklearn.preprocessing import normalize
 from torch import nn
-from torch.amp import autocast # pyright: ignore[reportPrivateImportUsage] 
+from torch.amp import autocast # pyright: ignore[reportPrivateImportUsage]
 from torch.utils.data import Dataset, DataLoader
 from torchvision.transforms import v2
 from torchvision.io import decode_image
@@ -33,25 +33,25 @@ def verify_pytorch() -> None:
     try:
         import torch
         print(f"✓ PyTorch {torch.__version__} detected")
-        
+
         # Check all device options
         if torch.cuda.is_available():
             print(f"✓ CUDA GPU available: {torch.cuda.get_device_name(0)}")
-        
+
         if torch.backends.mps.is_available():
             print("✓ Apple Silicon (MPS) GPU available")
-        
+
         if not torch.cuda.is_available() and not torch.backends.mps.is_available():
             print("! No GPU acceleration available. Expect slow feature extraction.")
-            
+
         return None
-    
+
     except ImportError:
         print("✗ PyTorch not found!")
         print("See homepage for PyTorch installation instructions")
 
         return None
-    
+
 def get_best_device() -> LiteralString:
     """Select torch device based on expected performance.
     """
@@ -64,14 +64,14 @@ def get_best_device() -> LiteralString:
     else:
         device = "cpu"
         device_name = "CPU"
-    
+
     print(f"Using device: {device} ({device_name})")
     return device
 
 class FeatureExtractor:
     """Extract features from images.
 
-    Extract feature vectors for individual identification from images. 
+    Extract feature vectors for individual identification from images.
     Currently, FeatureExtractor only includes the AnyDorsal algorithm.
 
     Parameters
@@ -82,31 +82,31 @@ class FeatureExtractor:
         Device with which to extract the features. By default, the best
         device is chosen for the user (cuda, mps, or cpu)
     stochastic : boolean, optional
-        Currently unused. 
+        Currently unused.
 
     Examples
     --------
     For a complete working example with real images, see:
-    
+
     - [Tutorial](../tutorial.ipynb)
-    
+
     Basic usage pattern::
-    
+
         from pyseter.extract import FeatureExtractor
-        
+
         # Initialize extractor
         extractor = FeatureExtractor(batch_size=16)
-        
+
         # Extract features from all images
         features = extractor.extract('path/to/images/')
-        
+
         # Access individual image features
         img_features = features['my_image.jpg']
-    
+
 
     """
-    def __init__(self, batch_size: int, 
-                 device: Optional[str]=None, 
+    def __init__(self, batch_size: int,
+                 device: Optional[str]=None,
                  stochastic: bool=False,):
         self.batch_size = batch_size
         self.stochastic = stochastic
@@ -120,22 +120,22 @@ class FeatureExtractor:
     def extract(self, image_dir: str, bbox_csv: Optional[str]=None) -> Dict:
         """Extracts features from images.
 
-        Extracts feature vectors for every image in a directory with the 
-        AnyDorsal algorithm.  
+        Extracts feature vectors for every image in a directory with the
+        AnyDorsal algorithm.
 
         Parameters
         ----------
         image_dir : str
-            Directory of images to from which to extract features. Directory 
+            Directory of images to from which to extract features. Directory
             should be flat, in that there should not be subdirectories with images.
         bbox_csv : str
-            Optional path to csv file with bounding boxes for each image in the 
+            Optional path to csv file with bounding boxes for each image in the
             image_dir.
 
         Returns
         -------
         dict
-            A mapping image file names to the corresponding feature vector. 
+            A mapping image file names to the corresponding feature vector.
             The file names are represented as strings, while the feature vector.
             is a NumPy array. For example:
             ```
@@ -147,12 +147,12 @@ class FeatureExtractor:
         Raises
         ------
         OutOfMemoryError
-            The GPU has run out of memory. Try reducing your batch size, or 
+            The GPU has run out of memory. Try reducing your batch size, or
             reducing the file size of the images in the directory.
         """
         print('Loading model...')
         model = self._get_model()
-        
+
         if not self.stochastic:
             model.eval()
 
@@ -162,7 +162,7 @@ class FeatureExtractor:
         features = self._extract_features(test_dataloader, model)
 
         return features
-    
+
     def _get_model(self):
         """Build the model from the checkpoint.
         """
@@ -176,14 +176,14 @@ class FeatureExtractor:
         # input/output dimensions for the model
         efficientnet_out_dim = 5504
         happywhale_class_count = 15587
-        
+
         # Create the complete model
         model = ImageClassifier(
             backbone=backbone,
-            in_channels=efficientnet_out_dim,  
-            num_classes=happywhale_class_count 
+            in_channels=efficientnet_out_dim,
+            num_classes=happywhale_class_count
         )
-        
+
         # download the model from huggingface
         model_path = hf_hub_download(
             repo_id=self.model_repo_id,
@@ -192,12 +192,12 @@ class FeatureExtractor:
 
         # Load the checkpoint
         checkpoint = torch.load(model_path, map_location=self.device)
-        
+
         # Process state dict
         state_dict = checkpoint["state_dict"]
         if 'head.compute_loss.margins' in state_dict:
             _ = state_dict.pop('head.compute_loss.margins')
-        
+
         # Adapt the state dict keys if needed
         new_state_dict = {}
         for k, v in state_dict.items():
@@ -219,22 +219,22 @@ class FeatureExtractor:
             else:
                 new_k = k
             new_state_dict[new_k] = v
-        
+
         # Load state dict with some flexibility for missing keys
         missing_keys, unexpected_keys = model.load_state_dict(new_state_dict, strict=False)
-        
+
         if missing_keys:
             print(f"Warning: Missing keys when loading pretrained weights: {missing_keys}")
         if unexpected_keys:
             print(f"Warning: Unexpected keys when loading pretrained weights: {unexpected_keys}")
-        
+
         # # Convert DropPath to inference mode if stochastic is enabled
         # if stochastic:
         #     model = convert_droppath_to_inference(model)
-        
+
         # Move model to device
         model.to(self.device)
-        
+
         return model
 
     def _extract_features(self, dataloader, model) -> dict:
@@ -249,19 +249,19 @@ class FeatureExtractor:
                     feature_vector = model(image, return_loss=False)
 
                 file_list.append(file)
-                feature_list.append(feature_vector.cpu().numpy())
+                feature_list.append(feature_vector.cpu().float().numpy())
 
         # Handle case with only one batch
         if isinstance(file_list[0], list) or isinstance(file_list[0], tuple):
             files = np.concatenate(file_list)
         else:
             files = np.array(file_list)
-            
+
         feats = np.vstack(feature_list)
 
         # Create dictionary mapping filenames to features
         feature_dict = dict(zip(files, feats))
-        
+
         return feature_dict
 
 class EfficientNetCustomBackbone(nn.Module):
@@ -275,20 +275,20 @@ class EfficientNetCustomBackbone(nn.Module):
             drop_path_rate=drop_path_rate,
         )
         self.with_cp = with_cp
-        
-        # Remove the classifier and pooling 
+
+        # Remove the classifier and pooling
         self.base_model.classifier = nn.Identity()
         self.base_model.global_pool = nn.Identity()
-        
+
     def forward(self, x):
         # We need to replicate the exact forward pass from your MMCLS TimmEfficientNet class
         # This extracts features before the pooling layer
-        
+
         # Apply stem
         x = self.base_model.conv_stem(x)
         x = self.base_model.bn1(x)
         # x = self.base_model.act1(x)
-        
+
         # Process through all blocks
         for block_idx, blocks in enumerate(self.base_model.blocks):
             for idx, block in enumerate(blocks):
@@ -296,17 +296,17 @@ class EfficientNetCustomBackbone(nn.Module):
                     x = cp.checkpoint(block, x)
                 else:
                     x = block(x)
-        
+
         # Final convolution and activation
         x = self.base_model.conv_head(x)
         x = self.base_model.bn2(x)
         # x = self.base_model.act2(x)
-        
+
         return x
 
 class NormLinear(nn.Linear):
     """Linear layer with optional feature and weight normalization."""
-    def __init__(self, in_features, out_features, bias=False, feature_norm=True, 
+    def __init__(self, in_features, out_features, bias=False, feature_norm=True,
                  weight_norm=True):
         super().__init__(in_features, out_features, bias=bias)
         self.weight_norm = weight_norm
@@ -327,10 +327,10 @@ class ImageClassifier(nn.Module):
     def __init__(self, backbone, in_channels=5504, num_classes=15587):
         super().__init__()
         self.backbone = backbone
-        
+
         # Global pooling layer
         self.neck = nn.AdaptiveAvgPool2d((1, 1))
-        
+
         # Classification head
         self.head = NormLinear(
             in_features=in_channels,
@@ -339,19 +339,19 @@ class ImageClassifier(nn.Module):
             feature_norm=True,
             weight_norm=True
         )
-        
+
     def forward(self, x, return_loss=True):
         # Extract features from backbone
         x = self.backbone(x)
-        
+
         # Apply global pooling and flatten
         x = self.neck(x)
         x = torch.flatten(x, 1)
-        
+
         # Return features if not computing loss
         if not return_loss:
             return F.normalize(x)
-        
+
         # Compute logits for training (not used in feature extraction)
         logits = self.head(x)
         return logits
@@ -392,7 +392,7 @@ class DorsalImageDataset(Dataset):
         if self.transform:
             image = self.transform(image)
         return filename, image
-    
+
 def get_test_data(directory, batch_size, bbox_csv=None):
     """Get the dataloader from a directory of images."""
     # RGB normalization values
@@ -410,14 +410,14 @@ def get_test_data(directory, batch_size, bbox_csv=None):
         v2.Normalize(rgb_mean.tolist(), rgb_std.tolist()),
     ])
 
-    test_data = DorsalImageDataset(directory, transform=data_transforms, 
+    test_data = DorsalImageDataset(directory, transform=data_transforms,
                                   bbox_csv=bbox_csv)
-    
-    dataloader = DataLoader(test_data, batch_size=batch_size, shuffle=False, 
+
+    dataloader = DataLoader(test_data, batch_size=batch_size, shuffle=False,
                            num_workers=2, pin_memory=True)
 
     return dataloader
-    
+
 def list_images(image_dir):
     """List all images in a directory."""
     images = []
@@ -436,7 +436,7 @@ def load_and_process_features(path, image_list, l2=True):
 
     if l2:
         feature_array = normalize(feature_array, axis=0)
-        
+
     return feature_array
 
 def load_all_features(feature_paths, image_list):
